backend/requirements.txt
# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Vector database and embeddings
chromadb>=0.4.0
openai>=1.3.0

# PDF and document processing
PyMuPDF>=1.23.0  # fitz for PDF parsing
python-docx>=1.1.0  # DOCX export

# Text processing and chunking
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
tiktoken>=0.5.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
